fusefs: more build fixes
* Fix printf format strings on 32-bit OSes
* Fix -Wclass-memaccess violation on GCC-8 caused by using memset on an object
  of non-trivial type.
* Fix memory leak in MockFS::init
* Fix -Wcast-align error on i386 in expect_readdir
* Fix some heterogenous comparison errors on 32-bit OSes.

Sponsored by:	The FreeBSD Foundation

+++ tests/sys/fs/fusefs/read.cc
@@ -486,9 +486,8 @@ TEST_F(ReadCacheable, mmap)
 	uint64_t ino = 42;
 	int fd;
 	ssize_t len;
-	ssize_t bufsize = strlen(CONTENTS);
+	size_t bufsize = strlen(CONTENTS);
 	void *p;
-	//char buf[bufsize];
 
 	len = getpagesize();
 
@@ -674,7 +673,7 @@ TEST_F(ReadCacheable, sendfile)
 	const char *CONTENTS = "abcdefgh";
 	uint64_t ino = 42;
 	int fd;
-	ssize_t bufsize = strlen(CONTENTS);
+	size_t bufsize = strlen(CONTENTS);
 	char buf[bufsize];
 	int sp[2];
 	off_t sbytes;
@@ -703,7 +702,8 @@ TEST_F(ReadCacheable, sendfile)
 
 	ASSERT_EQ(0, sendfile(fd, sp[1], 0, bufsize, NULL, &sbytes, 0))
 		<< strerror(errno);
-	ASSERT_EQ(bufsize, read(sp[0], buf, bufsize)) << strerror(errno);
+	ASSERT_EQ((ssize_t)bufsize, read(sp[0], buf, bufsize))
+		<< strerror(errno);
 	ASSERT_EQ(0, memcmp(buf, CONTENTS, bufsize));
 
 	close(sp[1]);
